moved Params and Locations to ArgumentCollections
