+++ src/test/java/org/broadinstitute/hellbender/tools/spark/sv/evidence/FindBreakpointEvidenceSparkUnitTest.java
@@ -31,7 +31,7 @@ public final class FindBreakpointEvidenceSparkUnitTest extends BaseTest {
     private final String asmQNamesFile = toolDir+"SVBreakpointsTest.assembly.qnames";
     private final String fastqFile = toolDir+"SVBreakpointsTest.assembly.";
 
-    private final FindBreakpointEvidenceSpark.Params params =
+    private final StructuralVariationDiscoveryArgumentCollection.FindBreakpointEvidenceSparkArgumentCollection.Params params =
             StructuralVariationDiscoveryArgumentCollection.FindBreakpointEvidenceSparkArgumentCollection.defaultParams;
     private final JavaSparkContext ctx = SparkContextFactory.getTestSparkContext();
     private final ReadsSparkSource readsSource = new ReadsSparkSource(ctx);
@@ -40,8 +40,8 @@ public final class FindBreakpointEvidenceSparkUnitTest extends BaseTest {
     private final JavaRDD<GATKRead> mappedReads = reads.filter(read -> !read.isUnmapped());
     private final ReadMetadata readMetadataExpected = new ReadMetadata(Collections.emptySet(), header, reads);
     private final Broadcast<ReadMetadata> broadcastMetadata = ctx.broadcast(readMetadataExpected);
-    private final FindBreakpointEvidenceSpark.Locations locations =
-        new FindBreakpointEvidenceSpark.Locations(null, null, null, null, null, null, null, null, null);
+    private final StructuralVariationDiscoveryArgumentCollection.FindBreakpointEvidenceSparkArgumentCollection.Locations locations =
+        new StructuralVariationDiscoveryArgumentCollection.FindBreakpointEvidenceSparkArgumentCollection.Locations(null, null, null, null, null, null, null, null, null);
     private final Set<String> expectedQNames = loadExpectedQNames(qNamesFile);
     private final Set<String> expectedAssemblyQNames = loadExpectedQNames(asmQNamesFile);
     private final List<SVInterval> expectedIntervalList = Arrays.asList(testIntervals);
